backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
html2image==2.0.4.3
Pillow==11.0.0
python-dotenv==1.0.0
fastmcp==0.9.0

